# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_transport_http_user.AuthResponse:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/internal_transport_http_user.UserResponse'
    type: object
  internal_transport_http_user.CreateUserDTO:
    properties:
      firstName:
        type: string
      lastName:
        type: string
      password:
        minLength: 8
        type: string
      patronymic:
        type: string
      username:
        type: string
    required:
    - firstName
    - lastName
    - password
    - username
    type: object
  internal_transport_http_user.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  internal_transport_http_user.LoginUserDTO:
    properties:
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  internal_transport_http_user.RefreshUserDTO:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  internal_transport_http_user.UserResponse:
    properties:
      createdAt:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      patronymic:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для дипломного проекта
  termsOfService: http://swagger.io/terms/
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для авторизации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_user.LoginUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_user.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_user.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http_user.ErrorResponse'
      summary: Аутентификация пользователя
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh Token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_user.RefreshUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_user.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_user.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http_user.ErrorResponse'
      summary: Обновить токены авторизации
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_user.CreateUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_transport_http_user.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_user.ErrorResponse'
      summary: Зарегистрировать новго пользователя
      tags:
      - Auth
  /user/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_user.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение информации о текущем пользователе
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"